grouped components in specific folders #refactor
