+++ public/mix-manifest.json
@@ -1,5 +1,5 @@
 {
-  "/js/react-panel.js": "/js/react-panel.df5753734a52a4a1fd6b.js",
-  "/css/app.css": "/css/app.62b4e583b1340c6c4214.css",
-  "/js/react-public.js": "/js/react-public.5cd80fa482160dbb47ca.js"
+  "/js/react-panel.js": "/js/react-panel.f8c3e2d975b646ebf86a.js",
+  "/css/app.css": "/css/app.72b19f93a096067d911d.css",
+  "/js/react-public.js": "/js/react-public.188e71de31c6038c1d98.js"
 }
