+++ resources/assets/js/components/color/ColorSidebar.js
@@ -5,10 +5,10 @@ import { bindActionCreators } from 'redux';
 import ReactSVG from 'react-svg';
 import CopyToClipboard from 'react-copy-to-clipboard';
 import NotificationSystem from 'react-notification-system';
-import { deleteFont } from '../reducers/Bridge/BridgeApiCalls';
-import { paramsChecker, isPublic } from '../helpers';
+import { deleteColor } from '../../reducers/Bridge/BridgeApiCalls';
+import { paramsChecker, isPublic } from '../../helpers';
 
-class FontSidebar extends Component {
+class ColorSidebar extends Component {
 
   constructor(props) {
     super(props);
@@ -22,20 +22,21 @@ class FontSidebar extends Component {
     this.openSettings = this.openSettings.bind(this);
     this.openPrimary = this.openPrimary.bind(this);
     this.addNotification = this.addNotification.bind(this);
-    this.deleteFont = this.deleteFont.bind(this);
-    this.weightToText = this.weightToText.bind(this);
+    this.deleteColor = this.deleteColor.bind(this);
   }
 
   componentDidMount() {
-    this.notificationSystem = this.refs.notificationSystem;
+      this.notificationSystem = this.refs.notificationSystem;
   }
 
-  addNotification(){
-    this.notificationSystem.addNotification({
-      message: 'Link copied to clipboard ',
-      level: 'success'
-    });
+  addNotification() {
+      debugger;
+        this.notificationSystem.addNotification({
+            message: 'Color copied to clipboard',
+            level: 'success'
+        });
   }
+
 
   openSettings() {
     this.setState((prevState, _) => {
@@ -49,45 +50,25 @@ class FontSidebar extends Component {
     })
   }
 
-  deleteFont() {
-    const { deleteFont, bridge, font, history } = this.props;
-    if(deleteFont){
+  deleteColor() {
+    const { deleteColor, bridge, color, history } = this.props;
+    if(deleteColor()){
       history.replace('/project/' + bridge.id);
-      deleteFont(bridge.id, font.id);
+      deleteColor(bridge.id, color.id);
     }
-  }
-
-  weightToText(weight) {
-    if(weight){
-      return weight.replace('100', "Thin ")
-        .replace('200', "Ultra Thin ")
-        .replace('300', "Light ")
-        .replace('400', "Regular ")
-        .replace('500', "Medium ")
-        .replace('600', "Semi Bold ")
-        .replace('700', "Bold ")
-        .replace('800', "Ultra Bold ")
-        .replace('900', "Black ");
-    }
-    return weight;
   }
 
   render() {
-
     const openSettings = this.openSettings;
     const openPrimary = this.openPrimary;
     const addNotification = this.addNotification;
-    const deleteFont = this.deleteFont;
-
-
+    const deleteColor = this.deleteColor;
     const isPub = isPublic();
 
-    const { margin, fontLink } = this.state;
+    const { margin } = this.state;
     const marginStyle = "-" + margin + "px";
 
-    const { font } = this.props;
-    console.log(font);
-    const weight = this.weightToText(font.variant.variant);
+    const { color } = this.props;
 
     let settings = null;
     if(!isPub){
@@ -95,20 +76,20 @@ class FontSidebar extends Component {
           <div className="settings">
               <div className="buttons">
                   <a className="button-outline-delete settings-button"
-                     onClick={deleteFont}>DELETE</a>
+                     onClick={deleteColor}>DELETE</a>
               </div>
           </div>
       );
     }
 
-    if(!font)
+    if(!color)
       return <div> </div>;
 
     return (
         <div className="bm-item-list sidebar">
             <div className="sidebar-padding">
                 <div className="header">
-                    <span>Use Font</span>
+                    <span>Use Color</span>
                 </div>
                 {settings}
             </div>
@@ -116,27 +97,15 @@ class FontSidebar extends Component {
                 <hr/>
             </div>
             <div className="sidebar-client sidebar-margin-top">
-                <div className="sidebar-section">
-                    <div className="title">
-                                <span>
-                                    SOURCE FILE
-                                </span>
-                        <a className="download-icon" href={font.variant.link}
-                           download={font.variant.font_family.family + " " + font.variant.variant}>
-                            <i className="fas fa-download"/>
-                        </a>
-                    </div>
-                </div>
-            </div>
-            <div className="sidebar-client sidebar-margin-top">
                 <div className="sidebar-section-url">
-                    <div className="sidebar-little-title">FONT URL</div>
-                    <div className="url url-darker">
+                    <div className="sidebar-little-title">HEX</div>
+                    <div className="url url-lighter">
                         <input type="text"
-                        defaultValue={`${font.variant.link}`.substring(0,59)}/>
+                               readOnly
+                               value={"#" + color.hex}/>
                         <span id="copy-to-clip">
                                         <CopyToClipboard
-                                            text={font.variant.link}
+                                            text={"#" + color.hex}
                                             className="clipboard" onCopy={() => {
                                             addNotification()
                                         }}>
@@ -149,67 +118,83 @@ class FontSidebar extends Component {
                 </div>
             </div>
             <div className="sidebar-client sidebar-margin-top">
-                <div>
-                    <div className="sidebar-section-img-info">
-                        <div className="sidebar-image-info">
-                            <div className="sidebar-little-title">FONT FAMILY</div>
-                            <div className="info info-darker">{font.variant.font_family.family}</div>
-                        </div>
-                        <div className="sidebar-image-info">
-                            <div className="sidebar-little-title">WEIGHT</div>
-                            <div className="info info-darker">{weight}</div>
-                        </div>
+                <div className="sidebar-section-url">
+                    <div className="sidebar-little-title">RGB</div>
+                    <div className="url url-darker">
+                        <input type="text"
+                               readOnly
+                               value={"rgb(" + color.rgb.replace(" ", ", ") + ")"}/>
+                        <span id="copy-to-clip">
+                                        <CopyToClipboard
+                                            text={"#" + color.hex}
+                                            className="clipboard" onCopy={() => {
+                                            addNotification()
+                                        }}>
+                                                <span id="icon">
+                                                    <i className="far fa-copy"/>
+                                                </span>
+                                        </CopyToClipboard>
+                                    </span>
                     </div>
                 </div>
             </div>
             <NotificationSystem ref="notificationSystem"/>
         </div>
     );
-
   }
-
 }
 
-FontSidebar.propTypes = {
-  font: PropTypes.shape({
+ColorSidebar.propTypes = {
+  color: PropTypes.shape({
     id: PropTypes.number
   })
 };
 
 const dispatchToProps = (dispatch) => {
   return bindActionCreators({
-    deleteFont
+    deleteColor
   }, dispatch)
 };
 
-export default connect(state => state, dispatchToProps)(FontSidebar);
-
+export default connect(null, dispatchToProps)(ColorSidebar);
 
 {/*<div className="font-sidebar">*/}
     {/*<div className="primary-view" style={{marginLeft: marginStyle}}>*/}
         {/*<div className="head">*/}
             {/*{ settingsButton }*/}
-            {/*<h3>Fonts</h3>*/}
+            {/*<h3>Color</h3>*/}
         {/*</div>*/}
         {/*<div className="content">*/}
             {/*<section>*/}
-                {/*<h4>Source File</h4><a className="button" href={font.variant.link} download={font.variant.font_family.family + " " + font.variant.variant}>Download</a>*/}
                 {/*<div className="clipboard_and_text">*/}
                     {/*<div>*/}
-                        {/*<CopyToClipboard text={font.variant.link} className="clipboard" onCopy={() => {addNotification()}}>*/}
+                        {/*<CopyToClipboard text={"#" + color.hex} className="clipboard" onCopy={() => {addNotification()}}>*/}
                     {/*<span><ReactSVG*/}
                         {/*path="/images/clipboard.svg"*/}
                     {/*/></span>*/}
                         {/*</CopyToClipboard>*/}
                     {/*</div>*/}
-                    {/*<p>{font.variant.link}</p>*/}
+                    {/*<p>{"#" + color.hex}</p>*/}
                 {/*</div>*/}
-            {/*</section>*/}
-            {/*<section>*/}
-                {/*<h4>Info</h4>*/}
-                {/*<div className="text">*/}
-                    {/*<p><span className="prefix">Font Family</span> <span className="primar">{font.variant.font_family.family}</span></p>*/}
-                    {/*<p><span className="prefix">Weight</span> <span className="primar">{weight}</span></p>*/}
+                {/*<div className="clipboard_and_text">*/}
+                    {/*<div>*/}
+                        {/*<CopyToClipboard text={"rgb(" + color.rgb.replace(" ", ", ") + ")"} className="clipboard" onCopy={() => {addNotification()}}>*/}
+                    {/*<span><ReactSVG*/}
+                        {/*path="/images/clipboard.svg"*/}
+                    {/*/></span>*/}
+                        {/*</CopyToClipboard>*/}
+                    {/*</div>*/}
+                    {/*<p>{"rgb(" + color.rgb.replace(" ", ", ") + ")"}</p>*/}
+                {/*</div>*/}
+                {/*<div className="clipboard_and_text">*/}
+                    {/*<div>*/}
+                        {/*<CopyToClipboard text={"#" + color.hex} className="clipboard" onCopy={() => {addNotification()}}>*/}
+                    {/*<span><ReactSVG*/}
+                        {/*path="/images/clipboard.svg"*/}
+                    {/*/></span>*/}
+                        {/*</CopyToClipboard>*/}
+                    {/*</div>*/}
+                    {/*<p>{"#" + color.hex}</p>*/}
                 {/*</div>*/}
             {/*</section>*/}
         {/*</div>*/}
@@ -221,15 +206,16 @@ export default connect(state => state, dispatchToProps)(FontSidebar);
                     {/*path="/images/close.svg"*/}
                 {/*/>*/}
             {/*</div>*/}
-            {/*<h3>Font Settings</h3>*/}
+            {/*<h3>Color Settings</h3>*/}
         {/*</div>*/}
         {/*<div className="content">*/}
             {/*<section>*/}
                 {/*<div className="text">*/}
-                    {/*<a className="button" onClick={deleteFont}>Delete</a>*/}
+                    {/*<a className="button" onClick={deleteColor}>Delete</a>*/}
                 {/*</div>*/}
             {/*</section>*/}
         {/*</div>*/}
     {/*</div>*/}
     {/*<NotificationSystem ref="notificationSystem" />*/}
 {/*</div>*/}
+
