+++ resources/assets/js/pages/Bridge.js
@@ -7,10 +7,10 @@ import { getSectionTypes } from '../reducers/SectionType/SectionTypeReducer';
 import { Helmet } from 'react-helmet';
 import DebounceInput from 'react-debounce-input';
 import { updateBridgeSlugRequest } from '../reducers/Bridge/BridgeApiCalls';
-import FontSection from '../components/FontSection';
-import ColorSection from '../components/ColorSection';
-import IconSection from '../components/IconSection';
-import ImageSection from '../components/ImageSection';
+import FontSection from '../components/font/FontSection';
+import ColorSection from '../components/color/ColorSection';
+import IconSection from '../components/icon/IconSection';
+import ImageSection from '../components/image/ImageSection';
 import { DragDropContext } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
 import { Link } from 'react-router-dom';
